Move getApplicationStatusGroup and getUnitStatusGroup to utils.
