+++ src/app/selectors.js
@@ -1,5 +1,7 @@
 import { createSelector } from "reselect";
 
+import { getApplicationStatusGroup, getUnitStatusGroup } from "./utils";
+
 // ---- Selectors for top level keys
 /**
   Fetches the model list from state.
@@ -139,7 +141,7 @@ const groupModelsByStatus = modelData => {
     let highestStatus = "running";
     Object.keys(model.applications).forEach(appName => {
       const app = model.applications[appName];
-      const appStatus = getApplicationStatusGroup(app);
+      const { status: appStatus } = getApplicationStatusGroup(app);
       if (statusOrder.indexOf(appStatus) > statusOrder.indexOf(highestStatus)) {
         highestStatus = appStatus;
       }
@@ -150,7 +152,7 @@ const groupModelsByStatus = modelData => {
       }
       Object.keys(app.units).forEach(unitId => {
         const unit = app.units[unitId];
-        const unitStatus = getUnitStatusGroup(unit);
+        const { status: unitStatus } = getUnitStatusGroup(unit);
         if (unitStatus === "blocked") {
           grouped.blocked.push(model);
           return;
@@ -161,33 +163,6 @@ const groupModelsByStatus = modelData => {
   }
   return grouped;
 };
-
-/**
-  Returns the string status for the application.
-  @param {Object} application The application to check the status of.
-  @returns {String} The status of the application.
-*/
-const getApplicationStatusGroup = application => {
-  // Possible "blocked" or error states in application statuses.
-  const blocked = ["blocked"];
-  // Possible "alert" states in application statuses.
-  const alert = ["unknown"];
-  const status = application.status.status;
-  if (blocked.includes(status)) {
-    return "blocked";
-  }
-  if (alert.includes(status)) {
-    return "alert";
-  }
-  return "running";
-};
-
-/**
-  Returns the string status for the unit.
-  @param {Object} unit The unit to check the status of.
-  @returns {String} The status of the unit.
-*/
-const getUnitStatusGroup = unit => {};
 
 /**
   Returns an object containing the grouped model status counts.
